pyproject.toml
[project]
name = "3d-bounding-box-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "lightning>=2.6.1",
    "numpy>=2.4.4",
    "onnx>=1.21.0",
    "onnxruntime>=1.24.4",
    "onnxscript>=0.6.2",
    "open3d>=0.19.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "pytorch-lightning>=2.6.1",
    "pyyaml>=6.0.3",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "wandb>=0.25.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"

[tool.uv]
index-strategy = "unsafe-best-match"
